Fix: Ensure docker-compose is restarted on docker restart
Previously restarting docker.service (e.g. on an update) caused
docker-compose to be shut down and not being restarted. This caused
services started by docker-compose not to be started and the system to
have no running containers. This fix chains docker-compose directly to
docker and therefore systemd will start docker-compose when docker comes
up again.

Signed-off-by: Knut Ahlers <knut@ahlers.me>